Combine the stores into one; separate action handlers to multiple files
