+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "charticulator",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "private": true,
   "author": {
     "name": "Donghao Ren",
