+++ src/app/stores/defaults.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
+
 import { Dataset, Specification, uniqueID } from "../../core";
 
 /** Create a default glyph */
