+++ src/app/stores/index.ts
@@ -1,13 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-export { BaseStore } from "../../core/store/base";
 
-export { MainStore } from "./main_store";
-export { DatasetStore } from "./dataset";
-export {
-  ChartStore,
-  Selection,
-  MarkSelection,
-  ChartElementSelection,
-  GlyphSelection
-} from "./chart";
+export { BaseStore } from "../../core/store/base";
+export { AppStore, AppStoreState } from "./app_store";
+export * from "./selection";
