+++ src/app/views/file_view/index.tsx
@@ -5,7 +5,7 @@ import * as R from "../../resources";
 
 import { AbstractBackend } from "../../backend/abstract";
 import { ErrorBoundary, SVGImageIcon } from "../../components";
-import { MainStore } from "../../stores";
+import { AppStore } from "../../stores";
 import { classNames } from "../../utils";
 import { FileViewExport } from "./export_view";
 import { FileViewNew } from "./new_view";
@@ -13,10 +13,10 @@ import { FileViewOpen } from "./open_view";
 import { FileViewSaveAs } from "./save_view";
 
 export class CurrentChartView extends React.PureComponent<
-  { store: MainStore },
+  { store: AppStore },
   { svgDataURL: string }
 > {
-  constructor(props: { store: MainStore }) {
+  constructor(props: { store: AppStore }) {
     super(props);
     this.state = {
       svgDataURL: null
@@ -39,7 +39,7 @@ export class CurrentChartView extends React.PureComponent<
 }
 
 export interface FileViewProps {
-  store: MainStore;
+  store: AppStore;
   backend: AbstractBackend;
   defaultTab?: string;
   onClose: () => void;
