+++ src/app/views/index.ts
@@ -7,8 +7,6 @@ export { DatasetView } from "./dataset/dataset_view";
 export { MarkEditorView } from "./canvas/mark_editor";
 export { ChartEditorView } from "./canvas/chart_editor";
 
-export { StatusBar } from "./status_bar";
-
 export { AttributePanel } from "./panels";
 
 export { FileView } from "./file_view";
